Fixed some defects in [13395]: added support for image maps. Refs #10998.
It seems that PlantUML only supports using stdin and stdout together, and there doesn't seem to be a way to generate the image map content when outputting to stdout. The documentation on image maps from PlantUML is almost non-existent, so it is hard to be sure. Therefore, we output the markup to a file, and pass the filepath of the markup file to PlantUML, outputting the png and cmapx to files in the same directory.

Thanks to Erwin Rademakers for coming up with this solution, and testing r13395.

git-svn-id: https://trac-hacks.org/svn/plantumlmacro/trunk@13397 7322e99d-02ea-0310-aa39-e9a107903beb